+++ libraries/extractor/src/main/java/androidx/media3/extractor/metadata/id3/TextInformationFrame.java
@@ -25,6 +25,7 @@ import androidx.media3.common.MediaMetadata;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
 import com.google.common.collect.ImmutableList;
+import com.google.common.primitives.Ints;
 import com.google.errorprone.annotations.InlineMe;
 import java.util.ArrayList;
 import java.util.List;
@@ -174,6 +175,18 @@ public final class TextInformationFrame extends Id3Frame {
       case "TXT":
       case "TEXT":
         builder.setWriter(values.get(0));
+        break;
+      case "TCON":
+        @Nullable Integer genreCode = Ints.tryParse(values.get(0));
+        if (genreCode == null) {
+          builder.setGenre(values.get(0));
+          break;
+        }
+        @Nullable String genre = Id3Util.resolveV1Genre(genreCode);
+        if (genre != null) {
+          builder.setGenre(genre);
+        }
+        // Don't set a numeric genre that we don't recognize.
         break;
       default:
         break;