+++ libraries/extractor/src/test/java/androidx/media3/extractor/metadata/id3/Id3UtilTest.java
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2020 The Android Open Source Project
+ * Copyright 2024 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package androidx.media3.extractor.mp4;
+package androidx.media3.extractor.metadata.id3;
 
 import static com.google.common.truth.Truth.assertThat;
 
@@ -21,13 +21,19 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-/** Test for {@link MetadataUtil}. */
 @RunWith(AndroidJUnit4.class)
-public final class MetadataUtilTest {
+public final class Id3UtilTest {
 
   @Test
-  public void standardGenre_length_matchesNumberOfId3Genres() {
-    // Check that we haven't forgotten a genre in the list.
-    assertThat(MetadataUtil.STANDARD_GENRES).hasLength(192);
+  public void expectedNumberOfV1Genres() {
+    for (int i = 0; i < 192; i++) {
+      assertThat(Id3Util.resolveV1Genre(i)).isNotNull();
+    }
+  }
+
+  @Test
+  public void unrecognizedV1Genre_returnsNull() {
+    assertThat(Id3Util.resolveV1Genre(-1)).isNull();
+    assertThat(Id3Util.resolveV1Genre(200)).isNull();
   }
 }